Create an empty account for new clients.

+++ src/bdd/java/steps/CreateAccountStepDefinitions.java
@@ -3,7 +3,7 @@ package steps;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import net.diegolemos.account.Account;
+import net.diegolemos.bankapp.account.Account;
 import net.diegolemos.bankapp.client.Client;
 import org.glassfish.jersey.jackson.JacksonFeature;
 
@@ -11,7 +11,7 @@ import javax.ws.rs.client.WebTarget;
 
 import static javax.ws.rs.client.ClientBuilder.newClient;
 import static javax.ws.rs.client.Entity.json;
-import static net.diegolemos.Main.BANK_APP;
+import static net.diegolemos.bankapp.Main.BANK_APP;
 import static net.diegolemos.bankapp.client.ClientBuilder.aClient;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertThat;
@@ -38,7 +38,7 @@ public class CreateAccountStepDefinitions {
     }
 
     @Then("^balance of his new account is (\\d+) EUR$")
-    public void balance_of_his_new_account_is_EUR(int amount) throws Throwable {
+    public void balance_of_his_new_account_is_EUR(double amount) throws Throwable {
         Account account = resource("account/" + username).request().get(Account.class);
         assertThat(account.getBalance(), equalTo(amount));
     }
